Tone down the comment about hook prePuller
I think the hook prePuller offers a lot of useful benefits
in cases where it helps (anytime someone is not using
the configurator). It allows us to merge changes anytime without
having to worry about the next user who logs in having to
wait too long, and the fact that the configurator does not
work with this is one of the reasons I stopped working on it
and instead started pursuing [z2jh-configurator](https://github.com/yuvipanda/z2jh-configurator).
It also very rarely actually causes us issues, even more so
because it runs at deploy time, not runtime.

So I've toned down the perceived negative comment here,
and instead tried to write *why* it is disabled here
specifically.

Personally, it also made me sad to see it described as
'large complexity' and 'small benefit', given:

1. I don't agree with that *value judgement* at all
2. I personally built most of it in z2jh, and it made
   me sad to see it described that way.